Pass measurement period (MP) values as strings and split useEffect for MP
To avoid complexities with TZ, this implementation
relies on Java's LocalDate to strip off time and TZ
data. The resulting date value is passed as `yyyy-MM-dd`.

TODO: Add support for time component (hh:mm:ss) and TZ's.
Users will want to test for events occuring at the edge
of the MP.